add support for specific cloud in headless
In order to kick off a non-interactive deployment
you now have to run the following:

conjure-up openstack to maas

The 'to maas' tells us that we already know which
cloud to pick and to just go ahead and deploy the
big software with no interruption.

Leaving off the 'to <cloud>' which initiate a interactive
tutorial based install.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>