refactor
Rework controllers to support a better mechanism for switching
between renderers.

Also cleaned up a lot of the code that used classes since for the
most part we don't need them as module namespacing gives us what
we need.

This also includes the latest WIP service deploy code as that is the
next step in the road to 2.0

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>