updates
fetching from the charmstore is now based on the conjure-* tags and
pulls in the necessary metadata for grabbing descriptions and our
whitelist/blacklist clouds.

Also create a general definitions file to pull from for keywords like
openstack. We dont really have a great way to grab a general summary
for keywords so this will do for now.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>