Add option editing to service walkthru view
Works, but UX is broken, most charms have too many options to show on
one screen and scrolling through them all is a pain.

Also pushes the 'next' button off the end, should probably use a Frame
to keep those around and then maybe add a tab loop to get between the
parts of the Frame.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>