dont assume unknown means active
this was a bad idea as all services start off as unknown so we weren't
actually waiting on the services to start properly. rather than work
around this in our code just file bugs against the charms that don't
use status-set to let us know their service is ready to go.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>