Improve finding existing controller & adding model
Finds existing controller even if no current model or controller is
selected. Doesn't login(), so doesn't require 'juju status' to be
working to find a controller - it won't be if no model is selected.

Backgrounds creating new model on found controller, so UI is more
responsive. It's on the juju queue, so deploys will wait for add-model
to be done.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>