Simplify deploy_service so we can wait on it
prev design did not return a future, and in no-revno cases, only
enqueued the deploy calls in the callback from get_info, which meant
that add_relations was called before the deploys were queued.

simplify design to avoid the callback for get_info and just wait on the
metadata controller's future in the deploy func, so we can always return
a future from deploy_service that is safe to wait on.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>