Do not load services at deploystatusview init
The status view will be created before we wait for the deploy futures,
which may be before the bootstrap is done. In this case, refresh_nodes
will call model_status, which tries to login and will then fail.

If we don't call refresh_nodes in init, then it will be called after the
future is done and we know that we can log in.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>