Fixes #174
The reason why we removed the pre-deploy in the first place is because
we couldn't guarantee that a bootstrap was completed before attempting
to login to the api server.

Fortunately, we've recently added async queues so this renders this
issue moot and we can associate our pre deploy tasks with the same async
queue as juju bootstrap. This lets us reliably run the predeploy tasks
after a bootstrap is complete and we can successfully query required
parameters like the provider type.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>