Bump rev in master
For now lets keep the next pre-release revision in master to be carried
over to the release branch during the time of making the release.

The workflow is:

Take current rev and bump to match that rev in the release branch. Once
a release is cut, bump the rev in the master branch as the next
prerelease revision.

So if we do a 2.0.0.6 release we would merge the code into the release
branch verifying both debian/changelog and conjure/__init__.py are the
proper release version. Then bump the rev in debian/changelog and
conjure/__init__.py in master to start the next development release.

A makefile target with these changes is in the works.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>