load metadata after spell is chosen
In keyword search mode, we have an app config with a bogus spell-dir, so
we need to wait until the variants view to create the metadata controller.

A simpler version of this fix would just always create it in
bundlereadme, but I'm keeping it in _start for the other case so that it
can be ready slightly sooner and avoid showing "Readme Loading" for
direct loaded spells.