remove superfluous async submit for add_machines
Unnecessary extra async queue entry caused a bug where services could be
added before the machines they needed.

_do_add_machines just called juju.add_machines, which also enqueues a
function to do the actual adding. It was possible for the user to
quickly deploy all the services in between enqueueing _do_add_machines
and when it was actually run, making the resulting queue out of order
and failing the deployment. Removing this unnecessary layer of
abstraction makes that impossible.

Fixes #273.

Also add add_machines call to TUI path.

Fixes #208.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>