bail on errors
Just a bare return is enough.
I'd previously removed return because it was returning the return value
of __handle_exception, which doesn't return anything explicitly, so that
was unclear.

However if we don't stop execution here, we will show the steps view and
not the error view with the exception.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>