bring spells into the project
With the upcoming release of snaps we are going to start keeping the
spells inside the project dir so that they are included with each snap
release. This makes it easier to distribute and with snappy's auto
upgrade feature and versioned data directories we can easily keep users
in sync with the latest 'curated' spells.

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>