add journald logging
Prior to setting the cachedir we relied on rsyslog for our logging, this
made it easier to define loggers outside of the running conjure-up
process. In addition to logging to a file this adds support for logging
directly to systemd journald. This makes it possible for us to keep all
logs relating to conjure-up and spells in a single location for parsing.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>