add journald logging (#398)
* add journald logging

Prior to setting the cachedir we relied on rsyslog for our logging, this
made it easier to define loggers outside of the running conjure-up
process. In addition to logging to a file this adds support for logging
directly to systemd journald. This makes it possible for us to keep all
logs relating to conjure-up and spells in a single location for parsing.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* fix travis build

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* another attempt to fix travis-ci

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* dont require systemd via pip

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* switch travis-ci to trusty

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* use syslog handler if available

We check that /dev/log exists and is a socket before attempting this as
to not break those running this under charmbox (docker)

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* revert travis-ci changes

dont use trusty for now

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>