Fixes #451
We only checked for if running in a gui because we were pulling a new
key into the action 'input' which is the text input box in the
additional input. However, in headless mode we that 'input' key doesn't
exist so we need to at very least check for default or just return an
empty string.

Driveby fixes status only, it was broken because of how we login to the
controllers now rather than switching to each one.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>