Always set series for services based on bundle
Previously, series was set only in deploy. This messed up metadata
lookup.

Earlier than that, we had been setting series based on a default of
'trusty' in the CharmStoreID class, which would be wrong in the case
where an ID didn't have the series and the bundle's series wasn't
'trusty'.

Fixes #458.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>