Fixes #464
This addresses the issue where the blacklist wasn't being properly
rendered in the UI. It also will hide choosing a public cloud if maas is
the only thing in the whitelist metadata. This should cover all
scenarios where either a whitelist or blacklist is defined and if
neither are found in the metadata.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>