Make sure deploy view is shown after newcloud
When creating a new cloud like maas and after hitting add credential
button the view would not change to the deploy screen. it was blocked on
waiting for post_bootstrap to finish. moving this into the handle
bootstrap function allowed us to finish setting up the deploy while the
bootstrap was happening in the background.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>