fix bootstrap wait view when importing lxd images
the way juju outputs the importing of lxd images assumes they are
working with a real tty. in our case we don't have this ability to trick
juju into thinkin our ui is a real tty. so we need to strip the control
characters from juju's debug output and sanitize it so it doesn't scroll
the text out out of the defined output in our bootstrap wait view.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>