Fix crash when creating new clouds
Store the cloud type early.
DeployController needs to know the cloud type during __init__, but the
bootstrap might not be done yet, so get_controller_info can't be used to
get the cloud type.

Instead store it as soon as we have chosen a cloud, and use that stored
value instead of inspecting juju via the CLI.

Signed-off-by: Michael McCracken <mike.mccracken@canonical.com>