Fixes #495
This handles multiple cases we've seen wrt LXD and network bridges. So we check
for various items like IPv6 enabled, if an actual bridge exists, and if the
current user is in the LXD group.

Additionaly, we've added some postinst scripts to create a new LXD bridge for
users who haven't set one up previously.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>