Allow for 'side-loading' additional charm options for applications
In some cases like OpenStack there is a single bundle for multiple substrates,
like MAAS and LXD (localhost). This usually requires alterations of some of the
charm options like LXD to point to no block-devices if inside a container, etc.

This allows us to create a 'bundle-custom.yaml' with a dictionary of

services:
  lxd:
    options:
       block-devices: None

And that we can merge into the charm stores original bundle for further
customization by the spell authors.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>