lxd version checking
Since LXD version switches away from semver on minor releases 2.10 as apposed to
2.10.0 we can't reliable check on the versioning. This isn't a huge deal since
we are always using the lxd versions from their stable ppa for now. Again this
isn't ideal and I would like a way to have more control over the versions of
lxd's we use but that just isn't possible at the point in time with snaps.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>