prep vsphere support, clean up schema credentials
Getting ready for vsphere support, there are a couple bugs I'd like to see fixed
prior to enabling vsphere in conjure-up (lp #1671650).

Also took time to rework the schema model, this makes it a bit easier to
reference additional properties of some of the input fields along with making
calls to those properties more readable.

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>