Hide empty log section ("Waiting") on interstitial
During JaaS controller registration and add-model if "Deploy All" is
selected quickly enough on the Deploy screen during the JaaS flow, the
"bootstrap" waiting interstitial is shown to inform the user that work
is being done.  But the bootstrap log file won't exist for those cases
like it will for the actual bootstrap process, so it's bad to show an
unchanging "Waiting" message on the left.