Improve can_sudo check
We should check immediately after authenticating sudo that we can indeed
perform a passwordless sudo.  This will catch early the case where auth
caching is disabled by setting timestamp_timeout=0 in sudoers (though it
will still fail).

Also, the -v flag does not work when actual passwordless sudo is enabled
for a user, so this drops it entirely in favor of executing /bin/true.