Use a auto generated name for credentials user key
First time users will enter credentials and our previous method was to use the
current_controller as our user key for selecting credentials during bootstrap.

This became unreliable as we were moving around ordering of views and controller
flow. Now we just auto generate the name and store that in our
app.current_credential variable and utilize it instead.

Fixes #907

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>