Prevent snap alias collisions
When using real snap aliases, the conjure-up snap will fail to install
if the juju snap is already installed, and it will block the juju snap
from installing if the conjure-up snap is already installed, unless the
user resolves the conflict manually (which may require uninstalling one
or the other snap).

This PR changes conjure-up to use a manual symlink which avoids both
conflict cases and automatically defers to the juju snap if present.
The only downside is that it doesn't show up in the `snap aliases` list.

Fixes #917