Do not filter out interfaces if run inside a container (#934)
* Do not filter out interfaces if run inside a container

If conjure-up is run inside a single lxd container then all network interfaces
are presented as virtual. The best approach to fixing this is to whitelist 'eth'
as this is the pattern for network device naming within containers. So for now
we check if virtual or if the device name doesn't start with eth.

Fixes #916

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>

* Fix logic

Fixes #915 as well

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>