Dont track certain exceptions
Handles ENOSPC cases as a no track exception so that we still show the error to
the user but dont treat it as an exception that is considered a bug.

For example, users system's with not enough disk space, nothing for us to fix
other than letting the user know that they've run out of disk space.

Fixes #959
Fixes #963

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>