Verify running user can access LXD daemon
Make sure the user is in the LXD group, if group doesn't exist or user is not in
the group give them a clear exception message on how to fix. Additonally, these
errors do not need to be tracked as it is understood that users with improper
permissions will never be able to access the lxd daemon bundled or not.

Fixes #973

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>