Pick unused port for bundled lxd to listen on
In our case we were setting port 12001 without realizing that NoMachine users
already allocate that port for remote desktop.

This isn't the greatest way but since we retry the port and setting a already
used port in LXD generates an error then this is relatively safe to assume that
an open port will be granted.

Fixes #975

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>