Fix JAAS registration (#984)
The calling protocol for `juju login` is different from `juju register`
which caused JAAS registration to break when it was switched in 5ab1c9b.

This also tries to login to JAAS once before prompting for credentials
in case there is an existing cookie / token.  This avoids prompting for
credentials when not needed and makes it clear when the cookie / token
is used.  It does, however, add an extra interstitial for the common
case.

Fixes #983