Small improvements to shutdown and error handling
The check in shutdown_watcher to not cancel itself wasn't correct, and
it wasn't giving cancelled tasks any loop cycles to be notified of the
cancellation, occasionally leading to spurious "task was pending"
errors.  Also expanded the catch-all for errors in the shutdown_watcher
to be more confident that the loop will stop, as well as ignoring
CancelledErrors in the exception handler, though they shouldn't ever
really reach it anyway.