Save VSphere regions in memory for new cloud
This fixes a NoneType error when adding a new VSphere cloud due to the
list of regions in memory being empty.

Also, cache the list of datacenters to save a call to the VSphere API,
and raise a more informative error if we can't get the info we need
about the datacenter.