Refactor how steps are run
Change how steps are run, and how app.env is populated for the steps, to
ensure a consistent environment for all steps.  This also makes it so
that all steps have access to all step data, in case they want to use it
for something (this enabled data to be prompted for steps such as
pre-deploy).  Also include JUJU_REGION in the env data.

Finally, fix one of the tests creating a spurious .local dir.