Fix subordinates not showing on Deploy Status
In #1042, placeholders were added to the Deploy Status screen to avoid
it being empty when first shown.  However, since the placeholders are
based on the expected number of units, which for subordinates is 0,
placeholders can not be created for subordinates, and the view didn't
account for them being added later.  They will now be added as they
appear in the model.