Fix detail log files not being written to
It seems that something with the snap confinement is causing passing
file handles to asyncio.subprocess to fail silently.  The files get
created but are left empty.  This doesn't happen when run outside of
the snap.  This gets around it by handling the file streaming in
explicit tasks that we control.  It also allows for some duplication
to be removed, so that's nice.

Fixes #1220