Show disabled clouds and why they're disabled
Hiding the clouds that are not availabe entirely can be confusing and
make it seem like we don't support all clouds.  Instead, this shows them
in a greyed-out state and provides an explanation as to why they're
disabled.  (I didn't include the specific reason, spell vs add-on, but I
don't think that's really necessary after all.)

Fixes #1222