Add network and storage check for LXD to cloud screen
To provide earlier and more detailed feedback on the reason for LXD not
working, this adds the checks for networking and storage to the monitor
task on the Cloud Selection screen.  If the networks or storage aren't
available, localhost will be disabled and hilighting it will provide a
specific message about why it is disabled.

With this, all of the LXD and schema error classes were refactored into
the errors module.

Fixes #1431