Adds support for bypassing juju and performing custom deploys
Reuses the same addons/steps/deploy/status code with a few branches based on if
juju is required or not.

Could use some more refactoring and building on this, but this works and is a
good start i think.

Need spell branch: https://github.com/conjure-up/spells/tree/microk8s-spell

Signed-off-by: Adam Stokes <battlemidget@users.noreply.github.com>