Fix case where no whitelisted clouds are enabled
If there are whitelisted clouds, the first enabled one should be
selected.  If none of the whitelisted clouds are enabled, the first one
in the list that's in the whitelist should be selected.  If none of the
whitelisted clouds are in the list, the first one in the list should be
selected.

Fixes #1487